Visual Studio 2012 C# AccessB debug
This is a Visual Studio 2012 project on C#, it work with an PIC18F2550
programmed with AccessB firmware.

This application was made to make test during the develop of this
project, the GUI are two parts, one with popups to read at least 10 SFR
values and another where the user can do the application under test,
then use the popups to select the SFR that want to check.